PWN-7021 - review fixex
